+++ client/client.go
@@ -45,17 +45,9 @@ func (c *Client) Post(ctx context.Context, operationName, query string, variable
 		option(&cc)
 	}
 
-	// PostMultipart send multipart form with files https://gqlgen.com/reference/file-upload/ https://github.com/jaydenseric/graphql-multipart-request-spec
-	req, err := NewMultipartRequest(ctx, cc.endpoint, operationName, query, variables)
+	req, err := NewRequest(ctx, cc.endpoint, operationName, query, variables)
 	if err != nil {
-		return fmt.Errorf("failed to create post multipart request: %w", err)
-	}
-
-	if req == nil {
-		req, err = NewRequest(ctx, cc.endpoint, operationName, query, variables)
-		if err != nil {
-			return fmt.Errorf("failed to create post request: %w", err)
-		}
+		return fmt.Errorf("failed to create post request: %w", err)
 	}
 
 	return cc.do(req, out)